--- src/components/Student/Dashboard/StudentUpdate.scss
@@ -1,11 +1,0 @@
-.update-weight {
-  background-color: #3f3f3f;
-  color: white;
-  border-radius: 8px;
-  margin-top: 2%;
-  padding: 1%;
-  font-family: "Raleway", sans-serif;
-  font-size: 1rem;
-  text-transform: uppercase;
-  font-weight: 700;
-}
